refactor(utils): extract usage pricing normalization
